requirements.txt
appdirs==1.4.3
apturl==0.5.2
arrow==1.2.2
asgiref==3.5.2
autopep8==1.6.0
Babel==2.10.1
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
binaryornot==0.4.4
blinker==1.4
Brlapi==0.7.0
certifi==2022.5.18.1
chardet==3.0.4
charset-normalizer==2.0.12
Click==7.0
colorama==0.4.3
command-not-found==0.3
cookiecutter==1.7.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.3
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==4.0.6
django-cors-headers==3.8.0
django-debug-toolbar==3.5.0
django-extra-views==0.13.0
django-haystack==3.2.1
django-localflavor==3.1
django-oscar==3.1
django-oscar-paypal==2.0.0
django-phonenumber-field==5.2.0
django-rest-knox==4.1.0
django-tables2==2.3.4
django-treebeard==4.4
django-widget-tweaks==1.4.12
djangorestframework==3.12.4
duplicity==0.8.12.0
entrypoints==0.3
factory-boy==3.2.1
Faker==13.12.1
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
gpg===1.13.1-unknown
httplib2==0.14.0
idna==3.3
importlib-metadata==1.5.0
Jinja2==3.1.1
jinja2-time==0.2.0
kazam==1.4.5
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libvirt-python==6.1.0
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==2.1.1
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
packaging==20.3
Paperboy==1.0.1
paramiko==2.6.0
pexpect==4.6.0
phonenumbers==8.12.49
Pillow==9.1.1
pipenv==2021.5.29
pipx==0.12.3.1
platformdirs==2.3.0
poyo==0.5.0
protobuf==3.6.1
purl==1.6
pycairo==1.16.2
pycodestyle==2.8.0
pycountry==22.3.5
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyparsing==2.4.6
PyQt5==5.14.1
pyRFC3339==1.1
pysolr==3.9.0
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-decouple==3.6
python-slugify==6.1.1
python-stdnum==1.17
pytz==2022.1
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.27.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
sip==4.19.21
six==1.16.0
sorl-thumbnail==12.6.3
sqlparse==0.4.2
systemd-python==234
text-unidecode==1.3
toml==0.10.2
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.9
usb-creator==0.3.7
virtualenv==20.8.1
virtualenv-clone==0.5.7
wadllib==1.3.3
wagtail-nesting-box==0.2.1
xkit==0.0.0
youtube-dl==2020.3.24
zipp==1.0.0
